+++ internal/service/user/client/client.go
@@ -3,12 +3,12 @@ package client
 import (
 	"github.com/spf13/viper"
 	"github.com/trainking/goboot/internal/pb"
-	"github.com/trainking/goboot/pkg/boot"
+	"github.com/trainking/goboot/pkg/service"
 )
 
 // NewUserService 创建UserServcie客户端
 func NewUserService(serviceName string, config *viper.Viper) (pb.UserServiceClient, error) {
-	conn, err := boot.NewGrpcClientConn(serviceName, config)
+	conn, err := service.NewGrpcClientConn(serviceName, config)
 	if err != nil {
 		return nil, err
 	}
