+++ pkg/boot/instance.go
@@ -1,10 +1,7 @@
 package boot
 
 import (
-	"os"
-	"os/signal"
 	"strconv"
-	"syscall"
 
 	"github.com/trainking/goboot/pkg/idgen"
 	"github.com/trainking/goboot/pkg/log"
@@ -52,30 +49,5 @@ func (b *BaseInstance) Init() error {
 
 	// 初始化ID生成器
 	idgen.InitNode(b.IntanceID)
-	return nil
-}
-
-// BootServe 启动服务
-func BootServe(instance Instance) error {
-	// 初始化
-	if err := instance.Init(); err != nil {
-		return err
-	}
-
-	// 2. 优雅退出
-	go func() {
-		exitC := make(chan os.Signal, 1)
-		signal.Notify(exitC, syscall.SIGINT, syscall.SIGTERM)
-
-		<-exitC
-		instance.Stop()
-		os.Exit(0)
-	}()
-
-	if err := instance.Start(); err != nil {
-		instance.Stop()
-		return err
-	}
-
 	return nil
 }
