+++ pkg/service/client.go
@@ -1,4 +1,4 @@
-package boot
+package service
 
 import (
 	"github.com/spf13/viper"
